fix(ripple): centers the md-fab ripple and updates the main template to correct ripple directive
style(ripple): slows down ripple movement and cleans up code

This slows down the ripple's movement towards the center to make it more obvious that the ripple is coming from the mouse cursor.  It also renames the diagonalWidth to rippleWidth to be more descriptive.

fix(nav): fix ink-ripple that was missed in previous commit