fixed sticky translate X
In the case were sticky elements had a left margin to the container, the sticky elements were being translated to the right just before the next element arrives. But the sticky element gets a margin-left, when translating element up from the top as it scrolls into view, it was also translated to the right. (I don't know if I am clear, put subheader elements in a scrolling container where they are not aligned to the left but centered)