fix(tabs): Fix tabs swipping when using JQuery
When using dots in JQuery events name, an event namespace is created and Event.type doesn't give the same name as using jqLite. JQuery removes the namespace from Event.type, but JQlite doesn't support namespace and keep the whole dotted value.

This cause an issue in tabs swipping, because onSwipe listener is checking Event.type property value.

For consistency with jqLite and jQuery, it's better to avoid checking the Event.type value, and use two distinct listeners for right and left swipping.