feat(select): BREAKING: allow md-input-container label and remove md-select-label
closes #2684, closes #1586, closes #3307

BREAKING: `md-select-label` is no longer uses. The select will now
always display the selected value as the text in the `md-option`. This
makes for a better UX for users, as otherwise it was possible to
confuse them.

`md-select` still supports using the `placeholder` attribute outside of
an `md-input-container` but the preferred way of using `md-select` is
now as follows:

```html
<md-input-container>
  <label>State</label>
  <md-select ng-model="selectedState">
    <md-option value="{{state}}" ng-repeat="state in states">
  </md-select>
</md-input-container>
```