ensure input focus doesn't trigger `handleQuery()` after selecting item on autocomplete with `minLength=0`
The onMouseup handler is fired after the user selects an item from the drop down before the click event is fired to actually select the item.
Focus performs a query if the min length is zero, in doing so it potentially replaces all the items and destroys their event handlers so the event is never fired.

by placing the call to focus on the nextTick, we eliminate this bug.