feat(contactChips): Add md-highlight-flags support.
Contact Chips currently allows the use of `md-highlight-text` to
match the search text. Users have requested we add
`md-highlight-flags` as well so that the highlighted text can
match the search field parameters.

Mainly, if a match is found with case-insensitive search, it
would be nice if the highlight text was also case-insensitive.

fixes #3182. closes #4278.