fix(mdBottomSheet): Enable scrolling on touch devices.
When `<md-content>` recently added the `-webkit-overflow-scrolling: touch`
CSS property, the `z-index` of the bottom sheet was placed into a
separate stacking context and did not respect the `z-index` of the
scroll mask.

Fix by ensuring the bottom sheets appends the scroll mask to the same
parent as the bottom sheet.