fix(fabSpeedDial): Make hovering an option via CSS.
Initially, the speed dial was designed to always open
when the user hovered over any portion of the speed
dial (including the area where the actions would
eventually appear). However, this made the speed dial
unusable on mobile (and sometimes unusable on desktop)
because it disallowed the user from pressing anything
underneath the actions.

Add the `md-hover-full` CSS class to allow developer
configuration of this behavior.

Ensure fabToolbar also works this way, and fix jumpy
animation.

Also updated the docs/demos and made the demos easier
to use on mobile.

BREAKING CHANGE - The fabSpeedDial no longer automatically
opens when hovering over the invisible actions. Add the
`md-hover-full` class to enable this interaction.

Fixes #4259.