fix(mdSelect): Do not close on scroll
Do not close the md-select when scrolled and mouseup event is above the md-options.
Remember the mouse x-position of the mousedown event.
Take the remembered x-position instead of the mouseup-x-position to check if the user has scrolled or not.
Do not close the select menu when scrolling is detected.

See issue #4707 and #4078