fix(inputs): Fix floating label and char counter positions.
Recent updates to the `md-char-counter` class caused alignment
issues when the input was side-by-side with another input. The
underlying cause was that the other input uses `maring-top: auto`
instead of `margin-top: 0` which made the inputs "bottom aligned"
as the height of it's neighbors changed.

This fixes the CSS on the char counter, inputs and select to
better align and updates the inputs demo to show the select
next to an input with the char counter.

Also, I noticed a bug with the floating labels when using only
the placeholder, so this adds some additional tests and fixes
the bug.

Fixes #4872. Closes #4915.