fix(mdDialog): fixed alert and confim `md-transition-in` class attachment
Alert and confirm dialogs had in their template class attribute that was binded to `dialog.css`, in case the css property was undefined it caused the `md-transition-in` class to be applied but than deleted.

The solution was to change the usage to ng-class and let angular set the class.

fixes #4862