fix(mdInput): Fix md-maxlength validation
`mdMaxlength` currently doesn't validate properly on inputs with `type="number"`. When the `$modelValue` returns a number, its `length` property is `undefined`. This change parses the `$modelValue` to a string when it's appropriate to do so.

Fix md-maxlength validation to use $viewValue

Add a test to validate the changes to md-maxlength