Correct md-tooltip positioning when scrolled
There are several issues out there (e.g. #2406) that point to tooltip positioning when the page is scrolled and, on experiencing this issue myself, I found the comment in the getNearestContentElement that suggests that the enclosed loop is supposed to stop at md-content but does not. Adding that condition to the loop fixed the positioning issue for me.