(fix dialog) Make sure dialog only destroys itself.
Since destruction is deferred, there's a risk of another dialog
having been created by the time "destroy()" is called.
We therefore have to make sure that we only remove one from the stack
if it's still left in it (which would be the case if "destroy()"
gets called as a result of the scope being destroyed independently,
but wouldn't be the case if the dialog is destroyed as a result of
calling $mdDialog.hide()).