fix(tabs): labels with fraction CSS width disabling pagination
When the md-tab-label CSS width is calculated to a fraction numbem, its
value and element.offsetWidth don't match. When the paging container's
width is calculated, depending on how the offsetWidth of each label is
truncated/rounded, the resulting width is smaller then the actual sum of
all labels, and that breaks the line and forces the last label to go to
next line, and its offsetLeft to be zero. That behaviour disables the next
page button, causing the pagination to not work. The new width calculation
adds 1px to the offsetWidth to account for rounding/truncating errors,
resulting in a slightly bigger paging container.

Fixes #5794. Fixes #5770. Fixes #5692. Closes #5801.