fix(md-tabs): doesn't scroll when width of tab wider than width of md-tabs-canvas tag
When the width of the label is bigger than that of the canvas, the
pagination tries to move the viewport to the start of the tab when going
forward, and to its end when going backwards, and it gets stuck on that
position. I changed the behaviour so that:
- when paging forward, search for next tab after current offset, so it
  discards the current showing tab;
- when paging backward, if the tab is bigger than the current viewport,
  move the canvas to the start of the tab. Otherwise, keeps it on the end
  (current behaviour).

Fixes angular/material#5697.