feat(select): support for key up and down value change
Add new attribute to control arrow key functionality: `allow-keyboard-change`
 - Defaults to false.
 - Ignored if `multiple` is set.

The HTML select control allows the user to traverse through options with
the up and down arrow keys. This functionality allows for a better User
Experience for users that prefer to navigate a form without using a
mouse.

The existing keypress behavior for the `mdSelect` component deviates a bit from
the native control. The up and down arrow keys open the select menu. Once open,
the up and down arrow keys navigate through the values. To close the menu,
the user must press Space or Enter, or click the option. This adds more
work for a user that is performing data entry tasks.

The `allow-keyboard-change` attribute aims to align the UX for `mdSelect` with
the HTML select control (to increase data entry efficiency).