Incomplete installation instructions
I was a bit annoyed I had to hunt this down. I honestly was under the impression that adding the CSS file would be enough to get started just like many other UI frameworks. I was wrong; though not mentioning the module to add in the instructions those give that same impression.