fix(#6033): md-select won't open in debug mode
In my test, it seemed like selectScope was strictly equal to scope, also, I did not find any usage of isOpen so I remove it. May be shouldn't have done this last part but writing unused data in an unknown scope did not seems like a good idea to me.