fix(input): Spacing/padding issues.
The spacing and paddin of inputs was incorrect due to recent changes.

**Fixes**

 - Alter margins and padding

 - Automatically apply the error spacing only to block-level inupts

 - Add a new `md-error-spacing` class for inline inputs

**Other Additions**

 - Update all components with custom CSS for `md-inline-form` class

 - Allow usage of `md-error-spacing` at the form level to quickly
   apply the spacing to all `<md-input-container>`s

 - Add new `md-block-form` class for quickly applying the `md-block`
   formatting to all nested `<md-input-container>`s

Update docs/demos accordingly and add new inline demo.

BREAKING CHANGE

We no longer automatically add bottom spacing to all instance of the
`<md-input-container>`. It is now applied only if you use the `md-block`
or `md-error-spacing` classes to allow for more compact form inputs.