fix(select): Position incorrect if selection scrolled.
If the currently selected object scrolled into view because it
was further down in the list, the position of the menu was
incorrect upon opening because the `scrollTop` of the `contentNode`
was always `0` due to the ARIA changes which set it to
`display: none` upon closing.

Fix by modifying the display to `block` when we run the `isScrollable`
check and back afterward (usually `display: none` but we account for
other options).

Also fixes a slight `2px` positioning issue which made the menu not
perfectly line up with the underlying selection.

Fixes #6190.

  Closes #6354