refactor: Identify private CSS classes.
Many components use classes that should be noted as private/internal
classes. Update components accordingly by prefixing classes with
`_md-`.

Refactor the following components:

 * FAB SpeedDial/Toolbar
 * Backdrop
 * Subheader
 * Chips
 * Content

Refactor the `md-visually-hidden` class to `_md-visually-hidden`
across multiple components.

Also make a slight change to icon docs for better readability.

Refactor `<md-content>`'s `autoScroll` class to `md-no-momentum`
for better clarity and to define it as a public API class and add
appropriate docs.

BREAKING CHANGE

The `autoScroll` class of the `<md-content>` directive has been
renamed to `md-no-momentum` to increase clarity and mark it as
an API-level class.

If you use the `autoScroll` class in your code, please update
it to the new name `md-no-momentum`.

References #6653.