fix(speedDial): Ensure scale animation actions are invisible when closed.
When using the `scale` animation, the FAB Speed Dial's action items could
sometimes appear as small dots on the screen even when the component was
closed.

This was due to an old bug where closing a dialog from a hidden action
item was throwing an issue. This has since been fixed so we can revert
this back to the original code that used `scale(0)` instead of
`scale(0.1)` which ensured that the dialog always had a button to animate
to when closing.

Also fix an issue with the `ng-hide` animation in the "More Options" demo.

Fixes #6344. Fixes #6670.

  Closes #6786