docs(select): remove fixed height for select demo
At the moment we are fixing the height, this will caught issues with the select because it provides the content and a margin of `20px 0 26px`,
plus the message which will be shown when the user selects one option. That mean's that won't fit in, and the clickable area will be shrinked.

Fixes #6943

Closes #6955