fix(autocomplete): fix not found template detection when element is hidden
When the element is hidden at compilation time through a ng-if directive for example.
The stored variable as currently used will be trashed, so it will be undefined.

So there a possibilities to store that state, using a variable in the directive (ex. hashmaps with ids) or storing it as an attribute / class.

Fixes #7035 Fixes #7142

  Closes #7042