fix(input): ngMessages height changing.
When using `ng-messages` inside of an input container, the height
of the messages will change when appearing, which pushes other
elements out of the way. This was a regression from the previous
behavior which allowed space for a one-line message before pushing
other content out of the way.

This was caused by some changes to the icons inside of an input.

Fix by finding an alternate method to address the icon issue.

Fixes #7072.

Closes #7146