fix(FAB): Re-add line-height to fix ui-sref issues.
The `line-height` property of the FAB buttons was removed to fix
an issue with the tooltip demo. This had unintended consequences
on FABs that used `ui-sref` which made them `<a>` tags instead of
`<button>` tags.

The tooltip demo issue was caused by the parent `<md-content>`
applying it's padding to all of it's children.

Revert the line-height change and fix demos by wrapping child
content in a new `<div>` which accepts the extra padding.

Fixes #7087. Closes #7145