fix(tabs): remove inline css width if tabs should stretch
There is an issue when dynamically setting md-stretch-tabs. If shouldStretchTabs() evaluates to false at any time then a fixed width is added and never removed.

This fix updates the logic in updatePagingWidth() so that the inline style
is always either updated or cleared from `<md-pagination-wrapper>`
depending on shoulStretchTabs().

Add tests for md-pagination-wrapper.