fix(input): Fix transition when switching tabs in Safari.
In Safari, if you select an input, switch tabs, then switch back,
the CSS animation to float the label does not fire properly causing
the label to sit on top on any text that you copy/paste into the
field.

Fix by delaying the `setFocus()` until the next tick.

Fixes #4203.