fix(list): multi-line inner-text-container should not overflow
At the moment in a multi line list, the text content will always break the layout when a text is overflowing.
- That's why the ellipsis indicator wasn't working too

This can be fixed by keeping the initial inner-list container in place, by ignoring overflow.
Now the list design won't break and the ellipsis will show as expected.

Fixes #7065