docs(autocomplete): better describe the `md-autofocus` attribute
The current description for the autocomplete attribute was really confusing, and the user was wondering why the autocomplete wasn't retrieving focus automatically.
But they didn't know about the other behaviour of the actual `md-autofocus` directive.
The autocomplete attribute "wrap" is just a little addition which focus the related input instead of the element.

Fixes #7283