fix(autocomplete): probably clear the autocomplete.
At the moment we calling the `select(-1)` function, but that will cause problems, because it will stuck.
There is no need of calling the `select(-1)` function, we should just clear the loading indicator on an synchronous query.
On an async query, the loading indicator will be set by default to false on completion.

This fixes the issue, which required two clicks on a list item after cleared the autocomplete through the button.

Fixes #7180