fix(dialog, bottom-sheet) fix focus trap for dialog when clicking outside and fix escape to close for dialog and bottom-sheet.
When `clickOutsideToClose` is disabled and you click outside, the focus will move to another element which is over of the actual interim element.
This will prevent the `escapeToClose` feature from work. The solution is, to make the backdrop or container focusable through mouse click, so the focus will stay inside the interim element.
This fixes also a issue with the focus trap of the dialog, because when `clickOutsideToClose` is disabled and you click outside, the focus will move outside of the focus trap. (Like to the buttons, which opened the dialog)

Fixes #7387

Closes #7415

# Conflicts:
#	src/components/bottomSheet/bottom-sheet.js
#	src/components/dialog/dialog.js