fix(input, mixin): fix input placeholder mixin and fix label theming.
At the moment the placeholder mixing isn't working, so the the placeholder is always using the browsers default color (`darkgrey`).
This commit fixes the invalid mixin for the placeholder, so the floating label and normal placeholder won't look different anymore.

Also removed the `text-shadow`, because in the specs is no inset shadow, this makes the label look weird.
So the labels will now look as same as in the specs, except that the `foreground-3` color is currently not as same as in the specs. (because it has a opacity of 26, but it should have 38)

Fixes #7429. Closes #7432