fix(list): fix overflow for ng-click and fix firefox click issues
I recently fixed a issue with the firefox click, and it worked perfectly. But in another commit later I unintentionally *removed* the firefox fix. So this fix reintroduces the fix and made it compatible with the other list fixes.

In additon to that, this commit fixes the text overflow for multiline items with ng-click.

Fixes #7490 Fixes #7499