fix(checkbox): pointer events disable ripple events too
We should not disable the ripple events by using `pointer-events: none`.
IE11 always focuses the div, instead of the `md-checkbox`, and that's causing issues with ngModel etc.

I good trick is, to redirect the focus, instead of preventing the whole focus / pointer events.
This won't interfere with the ripple, but it will fix the IE11 issue.

https://connect.microsoft.com/IE/feedback/details/1028411/

Fixes #7538