fix(datepicker): enable scrolling when scope destroyed.
Currently when we change the route, while the pane is open, the scroll mask won't be removed.
That's why we should re-enable scrolling on scope destruction.

In this case, we also have to check for the pane to be open, because otherwise we may accidentally remove another scroll mask.

Fixes #7542