fix(mdDialog): fix issue #7725
Fix the padding of custom dialogs and add title tag to seperate title and content.
Custom dialogs use a md-dialog-title tag to wrap the title.
Use the `md-dialog-content` class with in combination with the `md-not-last` class to remove the bottom padding,
e.g. to insert a table.

change your code from this:

```html
<md-dialog>
  <md-dialog-content>
    everything goes here
  </md-dialog-content>
</md-dialog>
```

to this:

```html
<md-dialog>
  <md-dialog-title>
    title goes here
  </md-dialog-title>
  <md-dialog-content>
    content goes here
  </md-dialog-content>
</md-dialog>
```