fix(sidenav): update position of sidenav and backdrop if scrolled.
Currently the sidenav is not animating when the parent is scrolled, this
was caused by the absolute position.

This commit, updates the position of the sidenav and backdrop when the
parent is scrolled.

It sets the position to the current scroll position and sets the height to
the parents `clientHeight`. Using a `100%` height, will impact the
performance heavily.

Fixes #5850.