fix(toolbar): remove transition duration for $ngAnimate
* Other components, like `md-toolbar` use a custom transition on the background-color, color etc, but the related CSS selector has a high specificity.

* So for example, when ng-hide is applied, then we set the transition to `none`.
But that's not working sometimes, because the both selectors have the same specificity.

To fix this, we can apply `!important` or increase the selector specificity by using a CSS hack.

https://www.w3.org/TR/CSS2/cascade.html#specificity

The toolbar selector has in our case a specificity of 20.
* .md-button.ng-hide // 0,0,2,0
* .md-toolbar-tools .md-button // 0,0,2,0

------
So this is actually a thing, which is caused by $ngAnimate, which is not
able to solve this issue.

We need to reset the transition duration, when $ngAnimate looks for it,
manually.

$ngAnimate Issue: https://github.com/angular/angular.js/issues/8224#issuecomment-49259798

Fixes #7929

Closes #8190