fix(toast, list): use a pseudo element to be able to use min-height.
* Currently all browsers, except IE11 are supporting flexbox's align-items property when having a `min-height`.
This bug is solved in the Edge Browser, but the fix will never land in IE11.

The vertical alignment bug for the list component was already solved by a previous commit, but the fix was only working for IE11 on Windows 8+.
IE11 on Windows 7- was not giving the pseudo element a `inline-block` position. (#7629)

Also this commit, removes the fixed height on the toast and uses the IE11 as well. (#8092)
So it's now possible, that the toast grows and the vertical alignment is also working.

Fixes #7629. Fixes #8092.