feat(dialog): allow to specify a content element
This allows developers, to specify a pre-rendered/compiled element, which will be not compiled each time a dialog opens.

It is possible to specify:
- A DOM Element
- A String as CSS Selector

When using an element, which is already present in the DOM, it will be anchored to the dialog, which means, that we temporary fetch it from the DOM into the dialog and move it back to its old DOM position upon close.

Closes #7566.

Closes #8491