fix/perf(tabs): Fix performance and visual issues.
Initially, the tabs used DOMSubtreeModified to listen for changes
and update the height, pagination and inkbar styles. At some point,
we used a different method to update the height, but we were still
watching all DOM changes and updating the pagination and inkbar
styles whenever the contents changed.

Fix this with the following two changes:

1. Move the DOM watching to a new md-tabs-dummy-wrapper directive
   which only watches for changes to the tab's labels (not the
   content too) so that the updates are called less frequently.

2. Replace DOMSubtreeModified watching with new MutationObservers
   which are more drastically more performant.

Additionally, some recent changes caused some bugs in the tabs by
using a cached version of the internal elements instead of calling
`getElements()` which searches the DOM and refreshes the cached
elements.

Fix by adding more calls to `getElements()` where appropriate to
properly update the cache after tabs may have changed.

Fixes #5681. References #4940, and might cause PR #8293 to need a
rebase if this goes in first.