fix(menu): Fix icon/text alignment in Firefox.
By default, Firefox provides focus border styling which was modifying
the layout of the button elements within a menu item. The original
fix was to use `display: flex` to fix this, however, Firefox currently
disregards the `display` property on the buttons and was thus causing
the text to be on a separate line from the text.

Fix by using `display: inline-block` for both `<a>` and `<button>`
elements and adding some special padding for the `<a>` tags to fix
the top alignment. Also apply special Firefox-specific styling to
remove the focus border/styling.

Fixes #8464.

Closes #8675