fix(content): Reduce iOS flicker when scrolling.
Often times when scrolling on an iOS device, certain elements on
the page would flicker. This was most noticeable with inputs, but
happens elsewhere too.

Add a new `md-no-flicker` CSS class which is automatically applied
to the `<md-content>` which adds a non-changing CSS transform to
the element which fixes the flicker.

Update docs to mention new class and workarounds if it causes
issues for developers.

Fixes #7078.

Closes #8680