fix(tooltip): fix regression on touch devices with a mouse
Reverts most of the changes from #8700 and switches to a simpler approach that uses a repeated `touchstart` to hide the tooltip, instead of `touchend`, which fires too early.
Also re-enables mouse events for all kinds of devices, in order to cover cases where the devices has both a touchscreen and a mouse.

Fixes #8710.