fix(site): Fix navigation toggle to properly open on page refresh.
Currently, the navigation groups on the left would have their icon
appear to be toggled open, but the actual content would be invisible
because the element's height was not being properly set.

Fix by setting the height in a `$mdUtil.nextTick()` so that it fires
at the appropriate time (and can measure the height).

Additionally, add some code to scroll the selected item into view so
that users know which page they are visiting.

Fixes #8841.