Update docs for reset of value
I spent a lot of time trying to figure out how to reset the displayed value of an autocomplete that I was (re)using in a drawer. I eventually found the answer in this issue: https://github.com/angular/material/issues/3084.

So this update attempts to prevent others from struggling with this issue, which to seems basic enough for it to be in the docs.