fix(select): Allow non-english characters for keyboard selection.
Our previous method for checking whether or not the typed character
matched the first letter of an `<md-option>` was constrained to the
letters `[a-zA-Z]`.

Fix by allowing more characters and checking the validity via the
`String.fromCharCode()` method.

Add tests for Spanish and Simplified Chinese, as well as not swallowing
certain keys like function keys (F3), meta, ctrl and comma (`,`).

Fixes #7730.