wip(tooltip): Tooltip transition to mdPanel
This is a WIP function that will transition the tooltip component to the mdPanel API.

Many of the core pieces of functionality of the original tooltip are still used due to the number of issues that they overcame.

The primary changes were to allow the mdPanel to position, animate, create, open and close the tooltip, instead of relying on proprietary logic within the tooltip code to do those things.

**Needs:**
 - Spec tests
 - Removal of the old tooltip component and renaming of this tooltip to `md-tooltip` instead of `md-tooltip-md-panel`