fix(input): Add missing Polyfill for ng1.3 and fix tests.
Angular 1.3 apparently did not include the `AnimateRunner.all()`
method. Grab it from latest 1.5.8 code and add to our 1.3 polyfills.

Additional, our $animateCss polyfill for 1.3 modifies the `to` and
`from` options that are passed in and sets them to `null`. This was
causing the tests to fail. Fix by keeping a cloned reference for
test purposes.