chore(changelog): update changelog task to latest package
Currently we use a modificated version of conventional-changelog by @robertmesserle.

It worked like a charm, but the official `conventional-changelog` package introduced some
changes. Like the `from` option (which was manually implemented by Rob)

Also the CI was not able to cache that specific package and always fetched it from GitHub.

---

* Added some nice log messages about the current fork points.
* The new changelog contains a link for the Github comparison.

Closes #8673.