fix(dialog): focus dialog element when no actions are set
By default the dialog searches for possible focus targets. Using the `md-autofocus` directive and querying for buttons inside of the `md-dialog-actions` element.
Some developers don't have any actions inside of the dialog, so there will be no focus target and Screenreaders are not able to detect the new change.

The dialog should focus the actual dialog element when no focus target is available.
Users are still able to disable the auto focus (as same as before) by using the `focusOnOpen` option.

> `mockElementFocus` now mocks the HTMLElement prototype instead of the jqLite prototype.
> This allows us to test focus on native HTML elements as well.

Fixes #9271.