fix(dialog): do not compile an empty element when using a content element
I recently noticed while playing with the pre-rendered dialog around, that despite of the pre-rendered dialog, the interim element triggers a compilation in the `$mdCompiler`.
It doesn't introduce any bugs, but it's unnecessary and could damage the performance.

Also moves the content element stuff into its own function, which makes the source more clean and allows us to maintain it easily.