feat(panel): Configuration ID for tracking
The user is now able to provide an ID as a property of the panel configuration object. When an ID is provided, the created panel is added to a tracked panels object. Any subsequent requests made to create a panel with that ID are ignored. This is useful in having the panel service not open multiple panels from the same user interaction when there is no backdrop and events are propagated.

Fixes #9356

Also, several syntax issues throughout the documentation/commenting have been corrected.

Fixes #9357

Ping ErinCoughlan