fix(datepicker): keyboard navigation not working if the user scrolls too much (#9302)
Fixes the datepicker's keyboard navigation not working if the user scrolls too much. This is due to the fact that after a month is destroyed, the browser automatically shifts focus to the body which prevents the calendar from picking up keyboard events. This change switches to listening for keyboard events on the body.

Fixes #9294.