docs(select): Add note about value/ng-value differences.
Some users expressed confusion about the fact that the `value`
attribute always evaluates to a string as opposed to the `ng-value`
attribute which allows any value.

Add a note to the docs to describe the differences and give an
example of how to properly use `ng-if` with the `value` attribute.

Fixes #9373.