fix(input): Ensure animated messages disappear.
After some recent changes to the input animations, the input could
get into a state where animated messages were still being shown
even if they were invalid.

The underlying issue was that the animation was not calling the
`done` function and Angular's animation framework was thus not
actually removing the element from the DOM, so any further message
animations found the old message in the DOM and attempted to show
it.

 - Ensure all input animations properly call the `.done()` method
   which is returned by `$animateCss()`.
 - Update tests to check for `.done()` calls and use a better
   method for flushing the animations.
 - Remove some old/commented code.

Fixes #9454.