fix(datepicker): improved overlay positioning (#9432)
The current approach for positioning the datepicker's calendar works by expanding the input container in order to fill a gap in the floating calendar pane and using negative margins to prevent the UI from jumping around. This comes with some drawbacks:
* It causes the input element to shift down slightly when the calendar gets opened.
* It can shift all of the elements surrounding the datepicker, despite the negative margins.
* It makes it hard to refactor the positioning.

This new approach solves all of the above-mentioned issues by using an element with a box-shadow to create the overlay, avoiding the need for expanding the container.

Referencing #9342.