perf(tooltip): reduce amount of event listeners on the window
Until now, each tooltip used to register 3 events on the window (scroll, resize and blur), even if it isn't being displayed. These can really add up on a page with a lot of tooltips. This change introduces a service that keeps track of the event handlers and only registers 1 global handler on the window for each event.

Referencing #9506.