chore(panel): fix failing unit test
When running the panel tests separately from all the others, the one about disabling scrolling was failing. This wasn't happening when running it together with all the other components, because some other component (not sure exactly which one, but it seemed like either dialog or autocomplete) wasn't cleaning up the DOM properly and was leaving a scroll mask lying around.

This change removes the DOM lookup for the scroll mask, because it looks like it got disabled in PR #8901, however the test didn't get updated.