fix(autofocus) properly parse attribute expression
The autofocus directive right now works with an empty value because jqLite just toggles the class if the `force` param is `undefined`.

This means if you change it several times to the same undefined value it does incorrectly update the autofocus class (which is required for the focus target)

Also the autofocus directive does now determine the class immediately and does not require a second digest (previously this could cause issues with the $mdCompiler linking).