fix(select): Fix duplicates in label.
Occasionally, `ng-repeat` appears to modify the DOM in such a way
that it allows duplicates of selected items. When this happens, the
`md-select` has issues displaying the proper label.

Add a quick filter to ensure there are no duplicate selections in
our labels.

NOTE: I was unable to reproduce this in a spec; it appears to be
related to the scope digets and never happens in the spec. I did
update one of the demos to do some manual testing and ensure this
fix works as expected, but it didn't have any real use inside of
the demo so I removed it before submission.

Fixes #9442.