﻿fix(autocomplete): dropdown shrink restore
The virtualRepeatContainer for the autocomplete dropdown determines the original dropdown size from the height of the container.

Commit b114302717341638bb627792c197ec3c43e069ff removed the fixed height of the dropdown selector, since we now have dynamic dropdown heights.

This does cause an issue for the virtualRepeatContainer, where it can't determine the `originalSize` of the dropdown.

Setting it to 100 percent, does properly expand the dropdown to the `max-height` of the JavaScript.

Fixes #9726.