fix(subheader): add accessibility support
Previously, due to the way the sticky headers work, lists with an `md-subheader` were being read out by screen readers as follows:
* Header 2
* Header 1
* Header 1
* Content 1
* Header 2
* Content 2

This can be confusing for users since it doesn't reflect the real structure of the page, as well as not reading out the headers as being headers.
These changes add the ARIA attributes that allow for the content to be read out as:
* Header 1
* Content 1
* Header 2
* Content 2

Fixes #9392.