fix(util): body overflow-x breaking disableScrollAround (#9864)
Currently the `disableScrollAround` function measures the body in order to determine whether to hide it's overflow. This measurement can get thrown off by the `html` node's scrollbar that gets added a few lines above. This change switches to adding the scrollbar to the `html` node after the body measurements are done.

Fixes #9860.