fix(autocomplete): Fix messages not appearing.
In certain scenarios, the `ng-messages` of an autocomplete would
fail to animate properly leaving them in a state where they were
in the DOM, but not visible to the user.

This was ultimately caused by the animations not properly retrieving
the messages element to animate, but was also tangentially caused by
the animations not properly calling the `done` callback upon failure.

There are three associated fixes:

 1. Log a warning if the messages element or it's children cannot be
    found.
 2. Ensure the `done` callback always fires even on failures.
 3. Ensure the `getMessagesElement()` method correctly returns when
    the element passed IS the messages element itself.

Fixes #9468.