feat(autocomplete): pass md-input-max/min-length to the input with non-floating label
Adds the md-input-minlength and md-input-maxlength attributes as ng-minlength and ng-maxlength to the input element of the autocomplete directive when a floating label is not used.
A use case : It would help in validating the input when the value of Search Text is being used if no matches are found.

Added Unit tests for md-input-maxlength and md-input-minlength