chore(components): use $onInit hook for binding assignments
A few components are using the `bindToController` option for the scope bindings.

Recently a Breaking Change commit landed in the AngularJS master branch, which changed the way the bindings are assigned.

The bindings are no longer pre-assigned in the constructors, and will be only set before the `$onInit` lifecycle hook.

Angular Material supports Angular versions which do not support lifecycle hooks and so we can't ensure that Snapshot and the older versions work together with `$onInit` approach.
---
* Adds the $onInit hook to all affected components and manually dispatches the $onInit hook for AngularJS 1.4 and older versions.

* Fixes two checks in the `datepicker`, which required the controller to have a binding set after $onInit (unnecessary here)