fix(mdTooltip): Tooltip position and watcher issue.
Fixes #10162 - When the position was changed dynamically the resulting
tooltip's panel would retain the position classes on it's panelEl.
Now when creating and recreating the tooltip with a different position
class due to a position change, the panel will grab that tracked panel
and update its configuration object so that the new position class can
be used.

Fixes #10157 - There was a `mdDirection` watcher that was deleted due
to a rebase/timing issue when moving to the `mdPanel` API. Reference
to previous PR: https://github.com/angular/material/pull/7822/files.