BugFix to $$interimElement
This affects mdDialog at the very least. I'm guessing this is somewhat of a typo, where someone forgot to add a paramter name to the resolve function. What happens is that another reponse value is captured, which is the promise that $mdDialog.show returns, so the result is we resolve a promise with itself, which Angular detects and and throws, so kudos to them.

Other weird things I noticed, but didn't look into :
1. The hideElement call on 529, doesn't call rejectAll on rejected promise. I would guess it should. 
2. if isCancelled is true, there is a call to rejectAll with the success result of the call of hideElement. Seems a bit weird.