fix(select): Accessibility fixes allowing screen readers (VoiceOver) to read list item position
1. Add role="presentation" to the container divs so assistive technology
(primarily VoiceOver) to accurately count and announce "item 3 of 7".
https://www.w3.org/TR/wai-aria/roles#presentation

2. Only add aria-owns when the md-select element does not contain its
menu element in the DOM. This means when the select is opened, as
opening the select means the element, previously nestled within
md-select, gets appended to the end of the DOM.

3. Dynamically hide md-select-value with aria-hidden to get an accurate
item count for the md-select. This can't just be hidden all the time
because VoiceOver does not allow linear navigation into selects with no
visible options.

4. Move autoFocus call to after the select menu element is fully rendered
in order to properly focus the first item in the select.

Addresses some points in #9908.